Homework 5. Переместил таск во фрагмент, сделал обработку отмены в onStop для фрагмента.
