+++ fragments/src/main/java/com/github/pashmentov96/fragments/ListPersonsFragment.java
@@ -1,16 +1,24 @@
 package com.github.pashmentov96.fragments;
 
+import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import java.util.List;
+
 public class ListPersonsFragment extends Fragment {
+
+    private static MyAsyncTask generatePersonsTask;
+    final String LOG = "MyLogs";
+    private PersonAdapter adapter;
 
     public static Fragment newInstance() {
         return new ListPersonsFragment();
@@ -31,9 +39,42 @@ public class ListPersonsFragment extends Fragment {
         recyclerView.setLayoutManager(new LinearLayoutManager(view.getContext()));
         recyclerView.setHasFixedSize(true);
 
-        final PersonAdapter adapter = new PersonAdapter();
+        adapter = new PersonAdapter();
         recyclerView.setAdapter(adapter);
 
-        adapter.setPersonList((ViewHolderListener) getActivity(), StorageOfPersons.getPersonList());
+        generatePersonsTask = new MyAsyncTask();
+        generatePersonsTask.execute();
+    }
+
+    class MyAsyncTask extends AsyncTask<Void, Void, List<Person>> {
+        @Override
+        protected List<Person> doInBackground(Void... voids) {
+            Log.d(LOG, "begin");
+            StorageOfPersons.generate();
+            Log.d(LOG, "end");
+            return StorageOfPersons.getPersonList();
+        }
+
+        @Override
+        protected void onPostExecute(List<Person> people) {
+            super.onPostExecute(people);
+            adapter.setPersonList((ViewHolderListener) getActivity(), people);
+            Log.d(LOG, "onPostExecute");
+        }
+
+        @Override
+        protected void onCancelled() {
+            super.onCancelled();
+            Log.d(LOG, "onCancelled");
+        }
+    }
+
+    @Override
+    public void onStop() {
+        super.onStop();
+        Log.d(LOG, "ListPersonsFragment: onStop");
+        if (generatePersonsTask != null) {
+            generatePersonsTask.cancel(true);
+        }
     }
 }
